+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- TypeScript toolchain for incremental adoption.
+
+### Changed
+
+- Updated [@adguard/filters-downloader] to v2.4.4.
+
+[Unreleased]: https://github.com/AdguardTeam/FiltersCompiler/compare/v3.2.9...HEAD
+
 ## [v3.2.9] - 2026-04-21
 
 ### Changed